GetSweepsWithSetting: Don't return invalid sweep numbers
Labnotebook entries before the first acquired sweep have NaN as sweep
number. This is not a valid sweep number, therefore GetSweepsWithSetting
should also not return that.

Bug present since forever.

+++ Packages/MIES/MIES_MiesUtilities.ipf
@@ -1656,6 +1656,12 @@ Function/WAVE GetSweepsWithSetting(labnotebookValues, setting)
 		Make/FREE/N=(DimSize(indizes, ROWS)) sweeps = labnotebookValues[indizes[p]][sweepCol][0]
 	endif
 
+	WAVE/Z cleanSweeps = ZapNaNs(sweeps)
+
+	if(!WaveExists(cleanSweeps))
+		return $""
+	endif
+
 	// @todo IP9: Make it threadsafe once FindDuplicates is threadsafe
 	return DeleteDuplicates(sweeps)
 End
